refactor: Separate unary-operator-related classes
- Plus/Minus
- Sine/Cosine/Tangent
