+++ Includes/Operators/UnaryOperator.hpp
@@ -26,56 +26,4 @@ public:
 	virtual std::string GetInfix(bool = false) const;
 };
 
-class UnaryMinus : public UnaryOperator
-{
-protected:
-	UnaryMinus(const UnaryMinus&);
-
-public:
-	UnaryMinus(ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Sine : public UnaryOperator
-{
-protected:
-	Sine(const Sine&);
-
-public:
-	Sine(ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Cosine : public UnaryOperator
-{
-protected:
-	Cosine(const Cosine&);
-
-public:
-	Cosine(ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Tangent : public UnaryOperator
-{
-protected:
-	Tangent(const Tangent&);
-
-public:
-	Tangent(ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
 #endif
